Rename Glitch64/main.h --> glitchmain.h for clarity.
Avoid mental association with Glide64/Main.cpp.
